adding the swagger to his config file
