+++ src/main.ts
@@ -1,19 +1,15 @@
 import { NestFactory } from '@nestjs/core';
-import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
 import { AppModule } from './app.module';
+import { SwaggerConfig } from './swaggerConfig'
 
 async function bootstrap() {
+
   const app = await NestFactory.create(AppModule);
 
-  const config = new DocumentBuilder()
-    .setTitle('Paim example')
-    .setDescription('The Pain API description')
-    .setVersion('1.0')
-    .addTag('Pain')
-    .build();
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('swagger', app, document);
+  SwaggerConfig(app);
 
   await app.listen(3000);
+
 }
+
 bootstrap();
